chore: move utils back into lib
