+++ src/components/bars/bar-list.tsx
@@ -9,7 +9,7 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card";
-import { cn } from "@/lib/utils/utils";
+import { cn } from "@/lib/utils";
 import Link from "next/link";
 
 type BarListProps = {
