+++ src/components/bars/bar-logo.tsx
@@ -1,5 +1,5 @@
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
-import { cn } from "@/lib/utils/utils";
+import { cn } from "@/lib/utils";
 
 const storageUrl = process.env.NEXT_PUBLIC_STORAGE_URL!;
