+++ src/components/ui/drawer.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { cn } from "@/lib/utils/utils";
+import { cn } from "@/lib/utils";
 import * as React from "react";
 import { Drawer as DrawerPrimitive } from "vaul";
