+++ src/components/ui/text-loop.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { cn } from "@/lib/utils/utils";
+import { cn } from "@/lib/utils";
 import {
   AnimatePresence,
   AnimatePresenceProps,
